Use int64 for bin indices

+++ python/tests/convert_test.py
@@ -97,8 +97,8 @@ def make_tof_binned_events():
         })
     return sc.bins(data=buffer,
                    dim='event',
-                   begin=sc.array(dims=['spectrum'], values=[0, 4]),
-                   end=sc.array(dims=['spectrum'], values=[4, 7]))
+                   begin=sc.array(dims=['spectrum'], values=[0, 4], dtype='int64'),
+                   end=sc.array(dims=['spectrum'], values=[4, 7], dtype='int64'))
 
 
 def make_count_density_variable(unit):
